cogs/moderation.py
"""Module dành cho các chức năng quản lý server (hiện trống)."""

from discord.ext import commands

class Moderation(commands.Cog):
    def __init__(self, bot):
        self.bot = bot

async def setup(bot):
    await bot.add_cog(Moderation(bot))